Add custom PartialEq implementation taking colour model into account
Rather than deriving `PartialEq` for `Luv` and `LCh` structs, add
custom implementation of that trait which takes into account
peculiarities of the colour model.

Firstly, if both colours have L* value equal zero they are considered
equal regardless of the other coordinates.  This is because colours
with L* of zero are both black.

Secondly, if two LCh colours with the same L* component have C*
component equal zero, they are considered equal regardless of their
hues.  This is because colours with C* equal zero are achromatic
shades of grey and their hue is undefined.

Thirdly, hues of two LCh colours are compared module τ.  This is
because hue is expressed as an angle around a circle and repeats every
τ radians so hue zero and hue τ are the same.